% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_XML2RDS.R
\name{Committee_XML2RDS}
\alias{Committee_XML2RDS}
\title{committee_XML2RDS}
\usage{
Committee_XML2RDS(filenameXLM = "",
                         filenameRDS = "",
                         WD_PathXML = "",
                         WD_PathRDS = "")
}
\arguments{
\item{filenameXLM:}{filename of the XML file}

\item{filenameRDS:}{filename of the RDS file}

\item{WD_PathXML:}{working directory where the XML file is located}

\item{WD_PathRDS:}{working directory where the RDS file will be saved}
}
\description{
This function transforms an XML file from a predefine working directory to an RDS file in a selected working directory
}
