% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_XML2RDS_ListFiles.R
\name{committee_XML2RDS_ListFiles}
\alias{committee_XML2RDS_ListFiles}
\title{committee_XML2RDS_ListFiles}
\usage{
committee_XML2RDS_ListFiles(WD_PathXML = "",
                            WD_PathRDS = "")
}
\arguments{
\item{WD_PathXML:}{working directory where the XML file is located}

\item{WD_PathRDS:}{working directory where the RDS file will be saved}
}
\description{
This function transforms all XML files from a predefine working directory to RDS files in a selected working directory
}
