% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committees_sittings_list.R
\docType{data}
\name{committees_sittings_list}
\alias{committees_sittings_list}
\title{Committees List}
\format{
A data frame with 14 787 rows and 5 variables
}
\source{
{Created for this package}
}
\usage{
committees_sittings_list
}
\description{
Contains information relevant to the committees dataset
}
\examples{
data(committees_sittings_list)
}
\keyword{datasets}
