% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_RDS_ListFiles.R
\name{compile_RDS_ListFiles}
\alias{compile_RDS_ListFiles}
\title{compile_RDS_ListFiles}
\usage{
compile_RDS_ListFiles(WD_PathRDS = "",
                      filename = "")
}
\arguments{
\item{WD_PathRDS:}{working directory where the RDS files are located}

\item{filename:}{output filename (singular file)}
}
\description{
This function compiles all RDS file in a selected working directory to a singular RDS file.
}
