% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee.R
\name{get_committee}
\alias{get_committee}
\title{get_committee}
\usage{
get_committee(Working_Directory = "~/demo_committee/xml/",
              StartDate = "2023-01-01",
              EndDate = "2023-06-30",
              Language = "en",
              CommitteeAbbrev = "INDU")
}
\arguments{
\item{Working_Directory:}{working directory where the XML files are going to be downloaded}

\item{StartDate:}{date to start gathering data (YYYY-MM-DD)}

\item{EndDate:}{date to stop gathering data (YYYY-MM-DD)}

\item{Language:}{English or French ("en", "fr")}

\item{CommitteeAbbrev:}{Committee abbreviation (4 letters)}
}
\description{
This function fetches committees data from the House of Commons of Canada website
}
\examples{
get_committee(Working_Directory = "~/demo_committee/xml/",
              StartDate = "2023-01-01",
              EndDate = "2023-06-30",
              Language = "en",
              CommitteeAbbrev = "INDU")

}
