% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hansard.R
\name{get_hansard}
\alias{get_hansard}
\title{get_hansard}
\usage{
get_hansard(Working_Directory = "",
            StartDate = "",
            EndDate = "",
            Language = "")
}
\arguments{
\item{Working_Directory:}{working directory where the XML files are going to be downloaded}

\item{StartDate:}{date to start gathering data (YYYY-MM-DD)}

\item{EndDate:}{date to stop gathering data (YYYY-MM-DD)}

\item{Language:}{English or French ("en", "fr")}
}
\description{
This function fetches Hansard data from the House of Commons of Canada website
}
\examples{
get_hansard(Working_Directory = "~/",
            StartDate = "2023-06-01",
            EndDate = "2023-06-30",
            Language = "en")

}
