% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vote.R
\name{get_vote}
\alias{get_vote}
\title{get_vote}
\usage{
get_vote(Working_Directory = "",
                Language = "",
                ParliamentNumber = "",
                SessionNumber = "")
}
\arguments{
\item{Working_Directory:}{working directory where the XML files are being download}

\item{Language:}{French or English ("fr", "en")}

\item{ParliamentNumber:}{Parliament number}

\item{SessionNumber:}{Session number}
}
\description{
This gathers all the votes for a specific Parliament and Session into an XML file
}
