% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vote_info_ListFiles.R
\name{get_vote_info_ListFiles}
\alias{get_vote_info_ListFiles}
\title{get_vote_info_ListFiles}
\usage{
get_vote_info_ListFiles(Working_Directory_Vote,
                               Working_Directory_voteInfo,
                               Language)
}
\arguments{
\item{Working_Directory_Vote:}{working directory where the XML files are located}

\item{Working_Directory_voteInfo:}{working directory where the XML files for vote details are going to be downloaded}

\item{Language:}{French or English ("fr" or "en")}
}
\description{
Loops over the vote directory to download all vote details in a single command
}
