% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_sittings_list.R
\docType{data}
\name{hansard_sittings_list}
\alias{hansard_sittings_list}
\title{Hansard List}
\format{
A data frame with 2419 rows and 1 variables
}
\source{
{Created for this package}
}
\usage{
hansard_sittings_list
}
\description{
Contains information relevant to the Hansard dataset
}
\examples{
data(hansard_sittings_list)
}
\keyword{datasets}
