% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hansard_XML2RDS_ListFiles.R
\name{hansard_XML2RDS_ListFiles}
\alias{hansard_XML2RDS_ListFiles}
\title{hansard_XML2RDS_ListFiles}
\usage{
hansard_XML2RDS_ListFiles(WD_PathXML = "~/demo_hansard/xml/",
                          WD_PathRDS = "~/demo_hansard/rds/")
}
\arguments{
\item{WD_PathXML:}{working directory where the XML file is located}

\item{WD_PathRDS:}{working directory where the RDS file will be saved}
}
\description{
This function transforms all XML files from a predefine working directory to RDS files in a selected working directory
}
\examples{
hansard_XML2RDS_ListFiles(WD_PathXML = "~/demo_hansard/xml/",
                          WD_PathRDS = "~/demo_hansard/rds/")

}
