% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_cleaning.R
\name{vote_cleaning}
\alias{vote_cleaning}
\title{vote_cleaning}
\usage{
vote_cleaning(xml_database = "")
}
\arguments{
\item{xml_database:}{XML file}
}
\description{
Cleans the XML nodes in preparation to RDS conversion
}
