% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_info_cleaning.R
\name{vote_info_cleaning}
\alias{vote_info_cleaning}
\title{vote_info_cleaning}
\usage{
vote_info_cleaning(xml_database = "")
}
\arguments{
\item{xml_database:}{XML file}
}
\description{
Cleans the XML nodes in preparation to RDS conversion
}
