% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_info_XML2RDS.R
\name{vote_info_XML2RDS}
\alias{vote_info_XML2RDS}
\title{vote_info_XML2RDS}
\usage{
vote_info_XML2RDS(XML_WD = "~/demo_vote_info/xml/",
                  RDS_WD = "~/demo_vote_info/rds/")
}
\arguments{
\item{XML_WD:}{working directory where the XML files are located}

\item{RDS_WD:}{working directory where the RDS files are going to be exported}
}
\description{
Transform vote details list from XML files to RDS files
}
\examples{
vote_info_XML2RDS(XML_WD = "~/demo_vote_info/xml/",
                  RDS_WD = "~/demo_vote_info/rds/")

}
