% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_XML2RDS.R
\name{vote_XML2RDS}
\alias{vote_XML2RDS}
\title{vote_XML2RDS}
\usage{
vote_XML2RDS(XML_WD = "~/demo_vote/xml/",
             RDS_WD = "~/demo_vote/rds/")
}
\arguments{
\item{XML_WD:}{working directory where the XML files are located}

\item{RDS_WD:}{working directory where the RDS files are going to be exported}
}
\description{
Transform vote list from XML files to RDS files
}
\examples{
vote_XML2RDS(XML_WD = "~/demo_vote/xml/",
             RDS_WD = "~/demo_vote/rds/")

}
