% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vote_info.R
\name{get_vote_info}
\alias{get_vote_info}
\title{get_vote_info}
\usage{
get_vote_info(Working_Directory = "~/demo_vote_info/xml/",
              Language = "en",
              ParliamentNumber = "44",
              SessionNumber = "1",
              DecisionDivisionNumber = "407")
}
\arguments{
\item{Working_Directory:}{working directory where the XML files are being download}

\item{Language:}{French or English ("fr", "en")}

\item{ParliamentNumber:}{Parliament number}

\item{SessionNumber:}{Session number}

\item{DecisionDivisionNumber:}{vote number}
}
\description{
This gathers vote details for a specific vote into an XML file
}
\examples{
get_vote_info(Working_Directory = "~/demo_vote_info/xml/",
              Language = "en",
              ParliamentNumber = "44",
              SessionNumber = "1",
              DecisionDivisionNumber = "407")

}
